requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
colorama==0.4.6
decorator==5.2.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
executing==2.2.0
gunicorn==23.0.0
inflection==0.5.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.1.1
wcwidth==0.2.13
